refactor(daemon): split acp.ts + pi-rpc.ts into agent-protocol/ capability-barrel
Move acp.ts (1744) and pi-rpc.ts (684) into a single agent-protocol/
capability-barrel module: a core/ foundation kernel (shared JSON-line-stream
transport) plus per-protocol concern subdirs (acp/, pi-rpc/), each with a
barrel. The shared createJsonLineStream transport moves into core/, dissolving
the former pi-rpc -> acp edge into a clean star topology (allowedEdges: []).

Pure structural move: every function body byte-identical (verified via
line-multiset, 1960==1960, 0 added/0 dropped); public surface reproduced
exactly (10 names) through the root barrel. All importers repointed to
./agent-protocol/index.js (server.ts, connectionTest.ts, runtimes/defs/shared.ts
+ 5 test files).

Guard registration deferred (check-barrel-imports infra not on main yet).
