+++ apps/daemon/src/agent-protocol/acp/index.ts
@@ -1,0 +1,3 @@
+export { type AcpMcpServerInput, buildAcpSessionNewParams } from './session-params.js';
+export { type ModelOption, normalizeModels, detectAcpModels } from './models.js';
+export { attachAcpSession } from './session.js';
