+++ apps/daemon/src/agent-protocol/acp/models.ts
@@ -1,0 +1,236 @@
+import { spawn } from 'node:child_process';
+import { createJsonLineStream } from '../core/index.js';
+import type { JsonRpcId, JsonObject, TimerHandle } from './types.js';
+import { ACP_PROTOCOL_VERSION, DEFAULT_TIMEOUT_MS, MODEL_CONFIG_OPTION_IDS } from './constants.js';
+import { errorMessage, resolveAcpTimeoutMs, asObject } from './json.js';
+import { sendRpc, rpcErrorMessage } from './rpc.js';
+import { buildAcpSessionNewParams } from './session-params.js';
+
+export interface ModelOption {
+  id: string;
+  label: string;
+}
+export interface AcpModelConfigOption {
+  configId: string;
+  currentValue: string | null;
+  values: unknown[];
+}
+export interface DetectAcpModelsOptions {
+  bin: string;
+  args: string[];
+  cwd?: string;
+  env?: NodeJS.ProcessEnv;
+  timeoutMs?: number;
+  clientName?: string;
+  clientVersion?: string;
+  defaultModelOption?: ModelOption;
+}
+export function normalizeConfigOptionToken(value: unknown): string {
+  return typeof value === 'string'
+    ? value.trim().toLowerCase().replace(/[\s_-]+/g, '')
+    : '';
+}
+export function isModelConfigOption(option: JsonObject, configId: string): boolean {
+  const category = normalizeConfigOptionToken(option.category);
+  if (category === 'model') return true;
+  const id = normalizeConfigOptionToken(configId);
+  if (id === 'model') return true;
+  if (category) return false;
+  const name = normalizeConfigOptionToken(option.name);
+  return MODEL_CONFIG_OPTION_IDS.has(id) || name === 'model';
+}
+export function findModelConfigOption(configOptions: unknown): AcpModelConfigOption | null {
+  const options = Array.isArray(configOptions) ? configOptions : [];
+  for (const rawOption of options) {
+    const option = asObject(rawOption);
+    if (!option) continue;
+    const configId = typeof option.id === 'string' ? option.id.trim() : '';
+    if (!configId) continue;
+    const type = typeof option.type === 'string' ? option.type.trim() : '';
+    if (type && type !== 'select') continue;
+    if (!isModelConfigOption(option, configId)) continue;
+    const currentValue =
+      typeof option.currentValue === 'string' && option.currentValue.trim()
+        ? option.currentValue.trim()
+        : null;
+    return {
+      configId,
+      currentValue,
+      values: Array.isArray(option.options) ? option.options : [],
+    };
+  }
+  return null;
+}
+export function normalizeModelConfigOptions(
+  configOptions: unknown,
+  defaultModelOption: ModelOption,
+): { currentModelId: string | null; models: ModelOption[] } | null {
+  const modelConfig = findModelConfigOption(configOptions);
+  if (!modelConfig) return null;
+  const seen = new Set([defaultModelOption.id]);
+  const out = [defaultModelOption];
+  for (const rawValue of modelConfig.values) {
+    const value = asObject(rawValue);
+    if (!value) continue;
+    const id =
+      typeof value.value === 'string' && value.value.trim()
+        ? value.value.trim()
+        : typeof value.id === 'string'
+          ? value.id.trim()
+          : '';
+    if (!id || seen.has(id)) continue;
+    seen.add(id);
+    const name = typeof value.name === 'string' ? value.name.trim() : '';
+    const isCurrent = id === modelConfig.currentValue;
+    const labelBase = name && name !== id ? `${name} (${id})` : id;
+    out.push({ id, label: isCurrent ? `${labelBase} • current` : labelBase });
+  }
+  return { currentModelId: modelConfig.currentValue, models: out };
+}
+export function normalizeModels(
+  models: unknown,
+  defaultModelOption: ModelOption,
+  configOptions?: unknown,
+): ModelOption[] {
+  const configModels = normalizeModelConfigOptions(configOptions, defaultModelOption);
+  if (configModels && configModels.models.length > 1) {
+    return configModels.models;
+  }
+  const modelsObj = asObject(models);
+  const available = Array.isArray(modelsObj?.availableModels) ? modelsObj.availableModels : [];
+  const currentModelId =
+    typeof modelsObj?.currentModelId === 'string' ? modelsObj.currentModelId : null;
+  const seen = new Set([defaultModelOption.id]);
+  const out = [defaultModelOption];
+  for (const model of available) {
+    const id = typeof model?.modelId === 'string' ? model.modelId.trim() : '';
+    if (!id || seen.has(id)) continue;
+    seen.add(id);
+    const name = typeof model?.name === 'string' ? model.name.trim() : '';
+    const isCurrent = id === currentModelId;
+    const labelBase = name && name !== id ? `${name} (${id})` : id;
+    out.push({ id, label: isCurrent ? `${labelBase} • current` : labelBase });
+  }
+  return out.length > 1 || !configModels ? out : configModels.models;
+}
+export function modelSelectionErrorIsRecoverable(code: unknown): boolean {
+  return code === -32603 || code === -32602 || code === -32601 || code === -32002;
+}
+export function currentModelFromSessionResult(result: JsonObject): string | null {
+  const configCurrent = findModelConfigOption(result.configOptions)?.currentValue;
+  if (configCurrent) return configCurrent;
+  const models = asObject(result.models);
+  return typeof models?.currentModelId === 'string' && models.currentModelId.trim()
+    ? models.currentModelId.trim()
+    : null;
+}
+export async function detectAcpModels({
+  bin,
+  args,
+  cwd = process.cwd(),
+  env = process.env,
+  timeoutMs = DEFAULT_TIMEOUT_MS,
+  clientName = 'open-design-detect',
+  clientVersion = 'runtime-adapter',
+  defaultModelOption = { id: 'default', label: 'Default (CLI config)' },
+}: DetectAcpModelsOptions): Promise<ModelOption[]> {
+  const effectiveTimeoutMs = resolveAcpTimeoutMs(env, timeoutMs);
+  return await new Promise<ModelOption[]>((resolve, reject) => {
+    const child = spawn(bin, args, {
+      cwd,
+      stdio: ['pipe', 'pipe', 'pipe'],
+      env: { ...env },
+    });
+    child.stdout.setEncoding('utf8');
+    child.stderr.setEncoding('utf8');
+
+    let settled = false;
+    let stderrBuf = '';
+    let expectedId = 1;
+    let nextId = 2;
+
+    let timer: TimerHandle | null = null;
+    const finish = <T extends ModelOption[] | Error>(fn: (value: T) => void, value: T) => {
+      if (settled) return;
+      settled = true;
+      if (timer) clearTimeout(timer);
+      try {
+        child.stdin.end();
+      } catch {}
+      fn(value);
+    };
+
+    const fail = (message: string) => {
+      finish(reject, new Error(message));
+      if (!child.killed) child.kill('SIGTERM');
+    };
+
+    const writeRpc = (id: JsonRpcId, method: string, params: unknown) => {
+      try {
+        sendRpc(child.stdin, id, method, params);
+      } catch (err) {
+        fail(`stdin write failed: ${errorMessage(err)}`);
+      }
+    };
+
+    const sendSessionNew = () => {
+      expectedId = nextId;
+      writeRpc(nextId, 'session/new', buildAcpSessionNewParams(cwd));
+      nextId += 1;
+    };
+
+    const parser = createJsonLineStream((raw) => {
+      const obj = asObject(raw);
+      const error = asObject(obj?.error);
+      const result = asObject(obj?.result);
+      const rpcErr = rpcErrorMessage(raw);
+      if (rpcErr) {
+        // JSON-RPC -32603 "Internal error" during model detection:
+        // If this is for the current expected-id (initialize/session/new),
+        // it's a real probe failure — reject immediately.
+        // Otherwise it's cleanup noise — suppress it.
+        if (error?.code === -32603 && obj?.id !== expectedId) return;
+        fail(rpcErr);
+        return;
+      }
+      if (obj?.id !== expectedId || !result) return;
+      if (expectedId === 1) {
+        sendSessionNew();
+        return;
+      }
+      if (expectedId === 2) {
+        const models = normalizeModels(result.models, defaultModelOption, result.configOptions);
+        finish(resolve, models);
+        if (!child.killed) child.kill('SIGTERM');
+      }
+    });
+
+    child.stdout.on('data', (chunk) => parser.feed(chunk));
+    child.stdout.on('close', () => parser.flush());
+    child.stdin.on('error', (err) => fail(`stdin error: ${err.message}`));
+    child.stderr.on('data', (chunk) => {
+      stderrBuf = `${stderrBuf}${chunk}`.slice(-16_000);
+    });
+    child.on('error', (err) => fail(`spawn failed: ${err.message}`));
+    child.on('close', (code, signal) => {
+      parser.flush();
+      if (!settled) {
+        const errTail = stderrBuf.trim();
+        const suffix = errTail ? ` stderr=${errTail}` : '';
+        fail(`ACP model detection exited code=${code} signal=${signal ?? 'none'}${suffix}`);
+      }
+    });
+
+    if (effectiveTimeoutMs > 0) {
+      timer = setTimeout(() => {
+        fail(`ACP model detection timed out after ${effectiveTimeoutMs}ms`);
+      }, effectiveTimeoutMs);
+    }
+
+    writeRpc(1, 'initialize', {
+      protocolVersion: ACP_PROTOCOL_VERSION,
+      clientCapabilities: { terminal: false },
+      clientInfo: { name: clientName, version: clientVersion },
+    });
+  });
+}
