+++ apps/daemon/src/agent-protocol/acp/rpc.ts
@@ -1,0 +1,95 @@
+import type { JsonRpcId, RpcWritable } from './types.js';
+import { asObject } from './json.js';
+
+export function sendRpc(writable: RpcWritable, id: JsonRpcId, method: string, params: unknown): void {
+  writable.write(
+    `${JSON.stringify({ jsonrpc: '2.0', id, method, params })}\n`,
+  );
+}
+export function sendRpcResult(writable: RpcWritable, id: JsonRpcId, result: unknown): void {
+  writable.write(`${JSON.stringify({ jsonrpc: '2.0', id, result })}\n`);
+}
+export function isJsonRpcId(value: unknown): value is JsonRpcId {
+  return typeof value === 'number' || typeof value === 'string';
+}
+export function rpcErrorMessage(raw: unknown): string {
+  const obj = asObject(raw);
+  const error = asObject(obj?.error);
+  if (!obj || !error) {
+    return '';
+  }
+  const message =
+    typeof error.message === 'string'
+      ? error.message
+      : typeof error.code === 'number'
+        ? String(error.code)
+        : 'json-rpc error';
+  return typeof obj.id === 'number'
+    ? `json-rpc id ${obj.id}: ${message}`
+    : message;
+}
+export function rpcErrorData(raw: unknown): unknown {
+  const obj = asObject(raw);
+  const error = asObject(obj?.error);
+  return error && 'data' in error ? error.data : undefined;
+}
+export function rpcErrorRetryable(data: unknown): boolean | undefined {
+  const details = asObject(data);
+  return typeof details?.retryable === 'boolean' ? details.retryable : undefined;
+}
+export function promotedOpenCodeSessionErrorPayload(data: unknown, fallbackMessage: string) {
+  const details = asObject(data);
+  if (
+    details?.kind !== 'opencode_session_error' ||
+    details.source !== 'opencode' ||
+    details.code !== 'ROLE_MARKER_HALLUCINATION'
+  ) {
+    return null;
+  }
+  const message =
+    typeof details.message === 'string' && details.message.trim()
+      ? details.message.trim()
+      : fallbackMessage;
+  return {
+    message,
+    error: {
+      code: 'ROLE_MARKER_HALLUCINATION',
+      message,
+      retryable: typeof details.retryable === 'boolean' ? details.retryable : true,
+      details: {
+        ...details,
+        promoted_by: 'open_design_acp',
+      },
+    },
+  };
+}
+export interface FormattedUsage {
+  input_tokens?: number;
+  output_tokens?: number;
+  cached_read_tokens?: number;
+  thought_tokens?: number;
+  total_tokens?: number;
+}
+export function formatUsage(usage: unknown): FormattedUsage | null {
+  const src = asObject(usage);
+  if (!src) return null;
+  const out: FormattedUsage = {};
+  if (typeof src.inputTokens === 'number') out.input_tokens = src.inputTokens;
+  if (typeof src.outputTokens === 'number') out.output_tokens = src.outputTokens;
+  if (typeof src.cachedReadTokens === 'number') {
+    out.cached_read_tokens = src.cachedReadTokens;
+  }
+  if (typeof src.thoughtTokens === 'number') out.thought_tokens = src.thoughtTokens;
+  if (typeof src.totalTokens === 'number') out.total_tokens = src.totalTokens;
+  return Object.keys(out).length > 0 ? out : null;
+}
+export function choosePermissionOutcome(options: unknown): string | null {
+  const list = Array.isArray(options) ? options : [];
+  const approveForSession = list.find((option) => option?.optionId === 'approve_for_session');
+  if (approveForSession) return 'approve_for_session';
+  const allowAlways = list.find((option) => option?.kind === 'allow_always');
+  if (allowAlways?.optionId) return allowAlways.optionId;
+  const allowOnce = list.find((option) => option?.kind === 'allow_once');
+  if (allowOnce?.optionId) return allowOnce.optionId;
+  return null;
+}
