+++ apps/daemon/src/agent-protocol/acp/session-params.ts
@@ -1,0 +1,68 @@
+import path from 'node:path';
+
+export interface AcpMcpServerInput {
+  type?: unknown;
+  name?: unknown;
+  command?: unknown;
+  args?: unknown;
+  env?: unknown;
+}
+export interface AcpSessionOptions {
+  mcpServers?: AcpMcpServerInput[];
+  // How the `env` field of each mcpServer entry is shaped.
+  // `'array'` (default) → `[{name, value}]` (Hermes, Kimi, …).
+  // `'map'`   → `{"KEY": "val"}` (reasonix 1.x Go, standard MCP).
+  envFormat?: 'array' | 'map';
+}
+export function buildAcpSessionNewParams(cwd: string, { mcpServers, envFormat = 'array' }: AcpSessionOptions = {}) {
+  const servers = Array.isArray(mcpServers) ? mcpServers : [];
+  const wantsMap = envFormat === 'map';
+  return {
+    cwd: path.resolve(cwd),
+    // MCP is an optional compatibility layer. Default to no MCP servers so ACP
+    // agents can run through the skill + CLI path without MCP support. Do not
+    // auto-install or mutate user/global MCP config; callers must pass an
+    // explicit per-session MCP descriptor when a compatible agent supports it.
+    mcpServers: servers.map((s) => {
+      const rawEnv = s?.env;
+      // Already a plain object — pass through in map mode, convert to
+      // array in array mode (e.g. live-artifacts MCP from
+      // buildLiveArtifactsMcpServersForAgent which already respects
+      // acpMcpEnvFormat).
+      const isPlainObject =
+        rawEnv && typeof rawEnv === 'object' && !Array.isArray(rawEnv);
+      if (wantsMap && isPlainObject) {
+        return {
+          type: typeof s?.type === 'string' ? s.type : 'stdio',
+          name: typeof s?.name === 'string' ? s.name : '',
+          command: typeof s?.command === 'string' ? s.command : '',
+          args: Array.isArray(s?.args) ? s.args : [],
+          env: rawEnv,
+        };
+      }
+      const envArr = Array.isArray(rawEnv) ? rawEnv : [];
+      const env = wantsMap
+        ? Object.fromEntries(envArr.map((e: any) => [e?.name ?? '', e?.value ?? '']))
+        : isPlainObject
+          ? Object.entries(rawEnv as Record<string, string>).map(
+              ([name, value]) => ({ name, value }),
+            )
+          : envArr;
+      return {
+        type: typeof s?.type === 'string' ? s.type : 'stdio',
+        name: typeof s?.name === 'string' ? s.name : '',
+        command: typeof s?.command === 'string' ? s.command : '',
+        args: Array.isArray(s?.args) ? s.args : [],
+        env,
+      };
+    }),
+  };
+}
+export function buildPromptBlocks(prompt: string, imagePaths: string[]): Array<Record<string, string>> {
+  const blocks: Array<Record<string, string>> = [{ type: 'text', text: prompt }];
+  for (const imagePath of imagePaths) {
+    if (typeof imagePath !== 'string' || imagePath.trim().length === 0) continue;
+    blocks.push({ type: 'resource_link', uri: imagePath });
+  }
+  return blocks;
+}
