+++ apps/daemon/src/agent-protocol/core/index.ts
@@ -1,0 +1,1 @@
+export { createJsonLineStream } from './json-line-stream.js';
