+++ apps/daemon/src/agent-protocol/index.ts
@@ -1,0 +1,10 @@
+export { createJsonLineStream } from './core/index.js';
+export {
+  type AcpMcpServerInput,
+  type ModelOption,
+  buildAcpSessionNewParams,
+  normalizeModels,
+  detectAcpModels,
+  attachAcpSession,
+} from './acp/index.js';
+export { mapPiRpcEvent, attachPiRpcSession, parsePiModels } from './pi-rpc/index.js';
