+++ apps/daemon/src/agent-protocol/pi-rpc/index.ts
@@ -1,0 +1,3 @@
+export { mapPiRpcEvent } from './events.js';
+export { attachPiRpcSession } from './session.js';
+export { parsePiModels } from './models.js';
