+++ apps/daemon/src/agent-protocol/pi-rpc/models.ts
@@ -1,0 +1,33 @@
+
+
+export type PiModelOption = { id: string; label: string };
+export function parsePiModels(stdout: unknown): PiModelOption[] | null {
+  const lines = String(stdout || '')
+    .split('\n')
+    .map((l) => l.trim())
+    .filter((l) => l.length > 0 && !l.startsWith('#'));
+
+  if (lines.length === 0) return null;
+
+  const DEFAULT_MODEL_OPTION = { id: 'default', label: 'Default (CLI config)' };
+
+  // First line is the header; skip it.
+  const entries = [DEFAULT_MODEL_OPTION];
+  const seen = new Set(['default']);
+  for (let i = 1; i < lines.length; i++) {
+    const line = lines[i];
+    if (line === undefined) continue;
+    const parts = line.split(/\s+/);
+    if (parts.length < 2) continue;
+    const provider = parts[0];
+    const modelId = parts[1];
+    if (provider === undefined || modelId === undefined) continue;
+    // Skip duplicates (some providers list the same model under multiple names).
+    const fullId = `${provider}/${modelId}`;
+    if (seen.has(fullId)) continue;
+    seen.add(fullId);
+    entries.push({ id: fullId, label: fullId });
+  }
+
+  return entries.length > 1 ? entries : null;
+}
