+++ apps/daemon/src/runtimes/defs/shared.ts
@@ -1,5 +1,5 @@
-import { detectAcpModels } from '../../acp.js';
-import { parsePiModels } from '../../pi-rpc.js';
+import { detectAcpModels } from '../../agent-protocol/index.js';
+import { parsePiModels } from '../../agent-protocol/index.js';
 import { execAgentFile } from '../invocation.js';
 import { DEFAULT_MODEL_OPTION } from '../models.js';
 import type { RuntimeModelOption } from '../types.js';
